requirements.txt
streamlit
pillow
pytesseract
huggingface-hub
python-dotenv
